twitcord_Back/requirements/base.txt
#pytz==2019.3  # https://github.com/stub42/pytz
#python-slugify==4.0.0  # https://github.com/un33k/python-slugify
Pillow==7.0.0  # https://github.com/python-pillow/Pillow
#argon2-cffi==19.2.0  # https://github.com/hynek/argon2_cffi
#whitenoise==5.0.1  # https://github.com/evansd/whitenoise
#redis==3.4.1 # https://github.com/andymccurdy/redis-py

# Django
# ------------------------------------------------------------------------------
#django==2.2.10  # pyup: < 3.0  # https://www.djangoproject.com/
#django-environ==0.4.5  # https://github.com/joke2k/django-environ
#django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
#django-allauth==0.41.0  # https://github.com/pennersr/django-allauth
#django-crispy-forms==1.8.1  # https://github.com/django-crispy-forms/django-crispy-forms
#django-redis==4.11.0  # https://github.com/niwinz/django-redis
# Django REST Framework
#djangorestframework==3.11.0  # https://github.com/encode/django-rest-framework
#coreapi==2.3.3  # https://github.com/core-api/python-client


# Django/ReactJS integration
# ------------------------------------------------------------------------------
#django-cors-headers==3.2.1  # https://github.com/ottoyiu/django-cors-headers
#graphene-django==2.8.1 # http://docs.graphene-python.org/projects/django/en/latest/
#graphene-django-optimizer==0.6.1  # https://github.com/tfoxy/graphene-django-optimizer
#graphene-file-upload==1.2.2  # https://github.com/lmcgartland/graphene-file-upload
#django-filter==2.2.0  # https://github.com/carltongibson/django-filter

Django==3.1.7
django-allauth==0.44.0
django-cors-headers==3.7.0
django-reset==0.2.0
django-rest==0.8.6
django-rest-auth==0.9.5
djangorestframework==3.12.2
oauth==1.0.1
#Pillow==5.1.0
psycopg2-binary==2.8.6
urllib3==1.26.4

# Storage
minio==7.0.2
channels==3.0.3
channels-redis==3.2.0
